chore: delete M68000 PRM PDF (replaced by markdown), grep port fixes, cleanups
- Remove M68000PRM.pdf — replaced by docs/references/m68000-prm/ markdown
- Grep port: fix d_type/DT_DIR handling, braceless if/else exit bug, recursive
  directory walking, updated test suite and binary
- Known pitfalls: add readdir() d_type and braceless if/else entries
- Port directory hygiene: add .claude/agent-memory/ to directory structure
- Shim: increase getdelim initial buffer from 512 to 4096 bytes
- Test harness: copy .dat files alongside .txt for FS-UAE tests
- Site packages: obfuscate uploader email in readme fields
- Remove stale port binaries (rebuilt in parallel session)

+++ ports/grep/ported/grep.h
@@ -48,7 +48,10 @@
 
 typedef struct {
 	size_t		 len;
-	long long	 line_no;
+	/* amiport perf: downgraded from long long — 2.1B lines is sufficient
+	 * and avoids 64-bit arithmetic on every line increment (68000 has no
+	 * 64-bit hardware: each long long add/compare is 4-6 instructions). */
+	long		 line_no;
 	off_t		 off;
 	char		*file;
 	char		*dat;
@@ -83,7 +86,9 @@ extern fastgrep_t *fg_pattern;
 extern regex_t	*r_pattern;
 
 /* For -m max-count */
-extern long long mcount, mlimit;
+/* amiport perf: downgraded from long long — -m values are bounded by INT_MAX
+ * (enforced by strtonum in main()); long is safe and avoids 64-bit ops. */
+extern long mcount, mlimit;
 
 /* For regex errors  */
 #define RE_ERROR_BUF 512
